refactor(api): artists + artist-notes → src/api modules (Effort 1)
Move src/hooks/queries/artists and its notes sub-slice into feature-sliced
src/api modules (src/api/artists and src/api/artist-notes) per ADR-0001.

Each feature gets a shared types.ts (entity Row type + query-key factory) and
one self-contained use* file per endpoint, with a per-file queryOptions factory
added for every query. Mutations keep useMutation with onSuccess/onError. No
behavior change; all consumers repointed and the cross-layer import from
src/api/sets/types.ts now targets @/api/artists/types.
