+++ src/api/artists/useArchiveArtist.ts
@@ -1,42 +1,43 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useToast } from "@/hooks/use-toast";
 import { supabase } from "@/integrations/supabase/client";
-import { artistsKeys } from "./useArtists";
+import { artistsKeys } from "./types";
 
-async function bulkArchiveArtists(artistIds: string[]) {
+// Mutation function
+async function archiveArtist(artistId: string) {
   const { error } = await supabase
     .from("artists")
     .update({ archived: true })
-    .in("id", artistIds);
+    .eq("id", artistId);
 
   if (error) {
-    console.error("Error bulk archiving artists:", error);
-    throw new Error("Failed to archive artists");
+    console.error("Error archiving artist:", error);
+    throw new Error("Failed to archive artist");
   }
 
-  return artistIds;
+  return true;
 }
 
-export function useBulkArchiveArtistsMutation() {
+// Hook
+export function useArchiveArtistMutation() {
   const queryClient = useQueryClient();
   const { toast } = useToast();
 
   return useMutation({
-    mutationFn: bulkArchiveArtists,
-    onSuccess: (results) => {
+    mutationFn: archiveArtist,
+    onSuccess: () => {
       queryClient.invalidateQueries({
         queryKey: artistsKeys.all,
       });
-
       toast({
-        title: "Bulk Archive Complete",
-        description: `Successfully archived ${results.length} artist(s).`,
+        title: "Success",
+        description: "Artist archived successfully",
       });
     },
     onError: (error) => {
       toast({
-        title: "Bulk Archive Error",
-        description: error?.message || "Failed to archive artists",
+        title: "Error",
+        description: error?.message || "Failed to archive artist",
         variant: "destructive",
       });
     },
