+++ src/api/artists/useArtists.ts
@@ -1,23 +1,8 @@
-import { useQuery } from "@tanstack/react-query";
+import { queryOptions, useQuery } from "@tanstack/react-query";
 import { supabase } from "@/integrations/supabase/client";
-import type { Database } from "@/integrations/supabase/types";
+import type { Artist } from "./types";
+import { artistsKeys } from "./types";
 
-export type Artist = Database["public"]["Tables"]["artists"]["Row"] & {
-  artist_music_genres: { music_genre_id: string }[] | null;
-  soundcloud_followers?: number;
-};
-
-// Query key factory
-export const artistsKeys = {
-  all: ["artists"] as const,
-  lists: () => [...artistsKeys.all, "list"] as const,
-  list: (filters?: unknown) => [...artistsKeys.lists(), filters] as const,
-  details: () => [...artistsKeys.all, "detail"] as const,
-  detail: (id: string) => [...artistsKeys.details(), id] as const,
-  bySlug: (slug: string) => [...artistsKeys.details(), "slug", slug] as const,
-};
-
-// Business logic function
 async function fetchArtists(): Promise<Artist[]> {
   const { data, error } = await supabase
     .from("artists")
@@ -58,10 +43,13 @@ async function fetchArtists(): Promise<Artist[]> {
   );
 }
 
-// Hook
-export function useArtistsQuery() {
-  return useQuery({
+export function artistsQuery() {
+  return queryOptions({
     queryKey: artistsKeys.lists(),
     queryFn: fetchArtists,
   });
+}
+
+export function useArtistsQuery() {
+  return useQuery(artistsQuery());
 }
