+++ src/pages/admin/ArtistsManagement/AddArtistDialog.tsx
@@ -24,8 +24,8 @@ import { Music } from "lucide-react";
 import { useAuth } from "@/contexts/AuthContext";
 import { useUserPermissionsQuery } from "@/api/auth/useUserPermissions";
 import { useGenresQuery } from "@/api/genres/useGenres";
-import { useCreateArtistMutation } from "@/hooks/queries/artists/useCreateArtist";
-import { useUpdateArtistMutation } from "@/hooks/queries/artists/useUpdateArtist";
+import { useCreateArtistMutation } from "@/api/artists/useCreateArtist";
+import { useUpdateArtistMutation } from "@/api/artists/useUpdateArtist";
 import { GenreMultiSelect } from "./GenreMultiSelect";
 import { FileUpload } from "@/components/ui/file-upload";
 import { uploadArtistLogo } from "@/services/storage";
