+++ src/pages/admin/ArtistsManagement/components/MergePreviewDialog.tsx
@@ -8,7 +8,7 @@ import {
 import { Button } from "@/components/ui/button";
 import { Label } from "@/components/ui/label";
 import { Loader2 } from "lucide-react";
-import type { Artist } from "@/hooks/queries/artists/useArtists";
+import type { Artist } from "@/api/artists/types";
 import { GenreBadge } from "@/components/GenreBadge";
 
 interface MergeChoices {
